создаю pointController и передаю всю логику event туда
